+++ src/Generator/TypeManagement/FieldDescriptor.cs
@@ -12,7 +12,7 @@ namespace IL2CS.Generator.TypeManagement
 	{
 		private static readonly Regex BackingFieldRegex = new("<(.+)>k__BackingField", RegexOptions.Compiled);
 
-		public FieldDescriptor(string name, TypeReference typeReference, FieldAttributes attrs, int offset)
+		public FieldDescriptor(string name, TypeReference typeReference, FieldAttributes attrs, ulong offset)
 		{
 			Name = BackingFieldRegex.Replace(name, match => match.Groups[1].Value);
 			// this is kinda evil, but it will make them consistent in name at least =)
@@ -26,7 +26,7 @@ namespace IL2CS.Generator.TypeManagement
 		public readonly string Name;
 		public readonly TypeReference Type;
 		public FieldAttributes Attributes;
-		public readonly int Offset;
+		public readonly ulong Offset;
 		public object DefaultValue;
 	}
 }
